**What this PR does / why we need it**:
- Split unicons out into a separate chunk
- Split `Chart` and `Geom` bizcharts components out using [`React.lazy`](https://reactjs.org/docs/code-splitting.html#reactlazy)

Lighthouse stats (with grafana server set to use http2):
Performance: 60 -> 62
First Contentful Paint: 2.5s -> 2.4s
Speed Index: 2.5s ->  2.4s
Largest Contentful Paint: 3.0s -> 2.8s
Time to Interactive: 2.6s -> 2.6s